pyproject.toml
[project]
name = "sort-my-roll"
version = "0.1.0"
description = "keep photo backups sorted by date-taken"
readme = "README.md"
authors = [
    { name = "Raphael Gabbarelli", email = "raphael@gabbarelli.it" }
]
requires-python = ">=3.13"

license = "MIT"
license-files = ["LICENSE"]

dependencies = [
    "click>=8.2.1",
    "pytest-mock>=3.14.1",
]

[project.scripts]
sort-my-roll = "sort_my_roll:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/raphaelgabbarelli/sort-my-roll"
